TDDExcercises/ValidationEngine/ValidationEngine.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Net.Mail;
using System.Text;
using System.Text.RegularExpressions;
using System.Threading.Tasks;

namespace ValidationEngine
{
    public class ValidationEngine
    {
        public bool IsValid { get; set; }
        public bool Validate(string email)
        {
            if (email == null)
            {
                return false;
            }
            var regex = new Regex(@"^([\w\.\-]+)@([\w\-]+)((\.(\w){2,3})+)$");
            var match = regex.Match(email);
            return match.Success;
        }
    }
}
